move exceptions from MitiResource to MitiService

+++ src/main/java/com/example/mitiappbackend/domain/miti/MitiService.java
@@ -23,6 +23,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.example.mitiappbackend.infrastructure.MitiCatchOnSameDayException;
+import com.example.mitiappbackend.infrastructure.MitiNotFoundException;
 
 @Service
 public class MitiService {
@@ -32,7 +33,23 @@ public class MitiService {
 
     @Transactional
     public void createMiti(Miti miti) throws MitiCatchOnSameDayException {
-        mitiRepository.createMiti(miti);
+        /*List<Miti> mitiList = mitiRepository.readMiti();
+        String mitiConcat = miti.getEmployee().getAbbreviation().getValue();
+        if (mitiList.contains(mitiConcat)) {
+            throw new MitiCatchOnSameDayException();
+        } else {
+            mitiRepository.createMiti(miti);
+        }*/
+        List<Miti> mitiRead = mitiRepository.readMiti();
+        List<String> mitiInfos = mitiRead.stream()
+            .map(Miti::catchMitiOnSameDay)
+            .toList();
+
+        if (mitiInfos.contains(miti.catchMitiOnSameDay())) {
+            throw new MitiCatchOnSameDayException();
+        } else {
+            mitiRepository.createMiti(miti);
+        }
     }
 
     @Transactional
@@ -41,17 +58,35 @@ public class MitiService {
     }
 
     @Transactional
-    public Miti readMitiByUuid(Long mitiId) {
+    public Miti readMitiByUuid(Long mitiId) throws MitiNotFoundException {
+        Miti mitiRead = mitiRepository.readMitiByUuid(mitiId);
+
+        if (mitiRead == null) {
+            throw new MitiNotFoundException(mitiId);
+        }
+
         return mitiRepository.readMitiByUuid(mitiId);
     }
 
     @Transactional
-    public void updateMitiByUuid(Long mitiId, Miti miti) {
+    public void updateMitiByUuid(Long mitiId, Miti miti) throws MitiNotFoundException {
+        Miti mitiUpdate = mitiRepository.readMitiByUuid(mitiId);
+
+        if (mitiUpdate == null) {
+            throw new MitiNotFoundException(mitiId);
+        }
+
         mitiRepository.updateMitiByUuid(mitiId, miti);
     }
 
     @Transactional
-    public void deleteMitiByUuid(Long mitiId) {
+    public void deleteMitiByUuid(Long mitiId) throws MitiNotFoundException {
+        Miti mitiDelete = mitiRepository.readMitiByUuid(mitiId);
+
+        if (mitiDelete == null) {
+            throw new MitiNotFoundException(mitiId);
+        }
+
         mitiRepository.deleteMitiByUuid(mitiId);
     }
 }
